Add first hacky version of `t`
I usually write notes for all my Jira tickets, so I wanted to automate
the creation of a new note. So far I was doing that manually *every
time*, which became super annoying and was actually stopping me from
taking notes sometimes, because the overhead of manually creating a new
note would be too large.

This is a first version, which is super hacky, but will hopefully serve
its purpose for now.